+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+
+- fix(registry): Replace the actual versionFilePath (#174)
 ## 0.17.0
 
 - feat(aws-lambda): Update the sentry release registry with AWS Lambda layer versions (#172)